!R (3DEngine) (Sandbox) Supported smaller heightmap unit size (down to 0.5m for now) for more detailed terrain !O (3DEngine) Optimized heightmap sector mesh generation: borders adjustment depending on current neighbor sectors lod is replaced with static safety borders Review: ID 66856 by Author: Vladimir Kajalin, Reviewer: Sebastien Laurent, Observer: Theodor Mader, Observer: Achim Lang, Observer: Nico Moss, Observer: Samuel Kahn, Observer: Luis Lairet  (Approved by theodor)
Copied from Perforce
 Change: 1536582

+++ Code/CryEngine/Cry3DEngine/3dEngine.h
@@ -427,9 +427,9 @@ public:
 	virtual void                     GenerateFarTrees(const SRenderingPassInfo& passInfo);
 	virtual float                    GetTerrainElevation(float x, float y, int nSID = GetDefSID());
 	virtual float                    GetTerrainElevation3D(Vec3 vPos);
-	virtual float                    GetTerrainZ(int x, int y);
-	virtual bool                     GetTerrainHole(int x, int y);
-	virtual int                      GetHeightMapUnitSize();
+  virtual float                    GetTerrainZ(float x, float y);
+  virtual bool                     GetTerrainHole(float x, float y);
+  virtual float                    GetHeightMapUnitSize();
 	virtual int                      GetTerrainSize();
 	virtual void                     SetSunDir(const Vec3& newSunOffset);
 	virtual Vec3                     GetSunDir() const;
