+++ Code/GameTemplates/cpp/Blank/Code/Components/Player.cpp
@@ -110,7 +110,7 @@ void CPlayerComponent::Revive()
 		// Position the player in the center of the map
 		const float heightOffset = 20.f;
 		float terrainCenter = gEnv->p3DEngine->GetTerrainSize() / 2.f;
-		float height = gEnv->p3DEngine->GetTerrainZ(static_cast<int>(terrainCenter), static_cast<int>(terrainCenter));
+		float height = gEnv->p3DEngine->GetTerrainZ(terrainCenter, terrainCenter);
 		Vec3 playerPosition = Vec3(terrainCenter, terrainCenter, height + heightOffset);
 
 		m_pEntity->SetWorldTM(Matrix34::Create(playerScale, playerRotation, playerPosition));
